review: Move default page selection out of terminal rendering
Review records, displayed-row IDs, and selectable-row calculations now live
in data.file_review. An output helper still chooses which pages to show
from requested pages and a previous line selection.

Commands need that decision before rendering or saving review state, and
choosing pages does not produce terminal output. Leaving the calculation in
output reverses the intended dependency.

This commit moves default page selection and its range-overlap helpers into
data.file_review, then updates commands to import page selection separately
from output.

The next commit will guard those command imports before the remaining saved
review state moves into the data package.
