+++ src/git_stage_batch/output/file_review_state_builder.py
@@ -3,7 +3,7 @@
 from __future__ import annotations
 
 from ..core.actionable_changes import ActionableSelectionReason
-from ..core.models import LineLevelChange, ReviewActionGroup
+from ..core.models import ReviewActionGroup
 from ..data.file_review.fingerprints import (
     compute_current_file_review_diff_fingerprint,
     fingerprint_selected_file_view,
@@ -14,7 +14,7 @@ from ..data.file_review.records import (
     FileReviewState,
     ReviewSource,
 )
-from ..data.file_review.pages import normalize_page_spec, parse_page_selection
+from ..data.file_review.pages import normalize_page_spec
 from ..data.selected_change.store import SelectedChangeKind
 from ..data.file_review.action_selections import (
     change_index_containing_review_display_ids,
@@ -23,7 +23,7 @@ from ..data.file_review.action_selections import (
     pages_containing_review_display_ids,
     selection_ids_for_display_ids,
 )
-from ..data.file_review.model import FileReviewModel, ReviewChange
+from ..data.file_review.model import FileReviewModel
 
 
 def _coerce_actionable_reason(reason: str) -> ActionableSelectionReason:
@@ -31,72 +31,6 @@ def _coerce_actionable_reason(reason: str) -> ActionableSelectionReason:
         return ActionableSelectionReason(reason)
     except ValueError:
         return ActionableSelectionReason.SIMPLE
-
-
-def resolve_default_review_pages(
-    model: FileReviewModel,
-    *,
-    requested_page_spec: str | None,
-    previous_selection: LineLevelChange | None = None,
-) -> tuple[int, ...]:
-    """Resolve explicit pages, selected-hunk anchor, or default page 1."""
-    page_count = len(model.pages)
-    if requested_page_spec is not None:
-        return parse_page_selection(
-            requested_page_spec,
-            page_count,
-            model.line_changes.path,
-        )
-    if page_count <= 1:
-        return (1,)
-    if previous_selection is not None and previous_selection.path == model.line_changes.path:
-        for change in model.changes:
-            if _change_overlaps_line_change(change, previous_selection):
-                return (change.first_page,)
-    return (1,)
-
-
-def _change_overlaps_line_change(change: ReviewChange, line_changes: LineLevelChange) -> bool:
-    old_numbers = [
-        line.old_line_number
-        for line in line_changes.lines
-        if line.kind != "+" and line.old_line_number is not None
-    ]
-    new_numbers = [
-        line.new_line_number
-        for line in line_changes.lines
-        if line.kind != "-" and line.new_line_number is not None
-    ]
-    return (
-        _ranges_overlap(
-            change.old_start,
-            change.old_end,
-            min(old_numbers, default=None),
-            max(old_numbers, default=None),
-        )
-        or _ranges_overlap(
-            change.new_start,
-            change.new_end,
-            min(new_numbers, default=None),
-            max(new_numbers, default=None),
-        )
-    )
-
-
-def _ranges_overlap(
-    left_start: int | None,
-    left_end: int | None,
-    right_start: int | None,
-    right_end: int | None,
-) -> bool:
-    if (
-        left_start is None
-        or left_end is None
-        or right_start is None
-        or right_end is None
-    ):
-        return False
-    return left_start <= right_end and right_start <= left_end
 
 
 def _supplemental_batch_review_selections(
